refactor: add common type in to generic type file
