+++ src/pages/interim/affiliate/_love-trading.tsx
@@ -3,27 +3,7 @@ import { LoveTradingComponent } from '../_love-trading'
 import { Flex } from 'components/containers'
 import { Header, Text, LiveChatLinkText } from 'components/elements'
 import { Localize } from 'components/localization'
-import { TString } from 'types/generics'
-
-type LeftCTASectionProps = {
-    header: TString
-    button_url?: string
-    button_text?: TString
-    hide_cta?: boolean
-    cta_props?: { is_white?: boolean }
-    custom_content?: React.ReactElement
-}
-type RightCTASectionProps = LeftCTASectionProps & {
-    button_props?: { type: 'dbot' | 'smart_trader' | 'mt5' }
-}
-
-type LoveTradingComponentProps = {
-    left: LeftCTASectionProps
-    right: RightCTASectionProps
-    image: string
-    bg_image?: string
-    bg_image_mobile?: string
-}
+import { LoveTradingComponentProps } from 'types/generics'
 
 const LoveTrading = () => {
     const content: LoveTradingComponentProps = {
