+++ src/pages/interim/dbot/_love-trading.tsx
@@ -1,27 +1,6 @@
 import React from 'react'
 import { LoveTradingComponent } from '../_love-trading'
-import { Localize } from 'components/localization'
-import { TString } from 'types/generics'
-
-type LeftCTASectionProps = {
-    header: TString
-    button_url?: string
-    button_text?: TString
-    hide_cta?: boolean
-    cta_props?: { is_white?: boolean }
-    custom_content?: React.ReactElement
-}
-type RightCTASectionProps = LeftCTASectionProps & {
-    button_props?: { type: 'dbot' | 'smart_trader' | 'mt5' }
-}
-
-type LoveTradingComponentProps = {
-    left: LeftCTASectionProps
-    right: RightCTASectionProps
-    image: string
-    bg_image?: string
-    bg_image_mobile?: string
-}
+import { LoveTradingComponentProps } from 'types/generics'
 
 const content: LoveTradingComponentProps = {
     left: {
