+++ src/pages/interim/dmt5/_love-trading.tsx
@@ -3,27 +3,7 @@ import { LoveTradingComponent } from '../_love-trading'
 import { Localize } from 'components/localization'
 import PlatformMt5 from 'images/svg/interim/bg-dmt5.svg'
 import PlatformMt5Mobile from 'images/common/interim/interim-mt5-mobile-bg.png'
-import { TString } from 'types/generics'
-
-type LeftCTASectionProps = {
-    header: TString
-    button_url?: string
-    button_text?: TString
-    hide_cta?: boolean
-    cta_props?: { is_white?: boolean }
-    custom_content?: React.ReactElement
-}
-type RightCTASectionProps = LeftCTASectionProps & {
-    button_props?: { type: 'dbot' | 'smart_trader' | 'mt5' }
-}
-
-type LoveTradingComponentProps = {
-    left: LeftCTASectionProps
-    right: RightCTASectionProps
-    image: string
-    bg_image?: string
-    bg_image_mobile?: string
-}
+import { LoveTradingComponentProps } from 'types/generics'
 
 const content: LoveTradingComponentProps = {
     left: {
