+++ src/types/generics.ts
@@ -118,3 +118,23 @@ export type TBrowserVersion = Partial<
         number
     >
 >
+
+type LeftCTASectionProps = {
+    header: TString
+    button_url?: string
+    button_text?: TString
+    hide_cta?: boolean
+    cta_props?: { is_white?: boolean }
+    custom_content?: React.ReactElement
+}
+type RightCTASectionProps = LeftCTASectionProps & {
+    button_props?: { type: 'dbot' | 'smart_trader' | 'mt5' }
+}
+
+export type LoveTradingComponentProps = {
+    left: LeftCTASectionProps
+    right: RightCTASectionProps
+    image: string
+    bg_image?: string
+    bg_image_mobile?: string
+}
